Handle impression listener specifically.
This allows us to have an onImpression callback property which is idiomatic for React.

Also because if the function changed in the config the new function wouldn't be called,
because JSON.stringify ignores functions. This is would be confusing for users.
Also the fact that we are using JSON.stringify should be transparent,
a hidden implementation detail, to users.